Services/FileService.cs
using placeholder.Models;

public class FileService
{
    private readonly FileRepository _fileRepository;

    public FileService(FileRepository fileRepository)
    {
        _fileRepository = fileRepository;
    }

    public async Task UploadFileAsync(IFormFile file)
    {
        // Logic to save the file
    }

    public async Task<List<FileModel>> GetAllFilesAsync()
    {
        return await _fileRepository.GetAllFilesAsync();
    }

    public async Task<FileModel> GetFileByIdAsync(int id)
    {
        return await _fileRepository.GetFileByIdAsync(id);
    }

    // Add other business logic methods here...
    public async Task AddFileAsync(FileModel file)
    {
        await _fileRepository.AddFileAsync(file);
    }
}
